Improve display hotplugging handling

+++ polybar/.config/polybar/launch.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+# Acquire mutex, exit if this script is already running.
+# This is to prevent errors caused by running this script multiple times
+# in rapid succession which happened often to me when using srandrd.
+if ! mkdir /tmp/polybar-launch.lock 2> /dev/null; then exit 1; fi
+
 # Terminate already running bar instances
 killall -q polybar
 
@@ -15,3 +20,6 @@ if [ "$(polybar -m | grep -w DP1)" ]; then
 fi
 
 echo "Bars launched..."
+
+# Free the mutex
+rm -r /tmp/polybar-launch.lock
